Fix bugs in and refactor TileMap and related
I changed TileMap’s poorly-named `isPositionInBounds` method to `clipPositionToBounds`, making it return `Unit` to follow the principle of CQS (Command-Query Separation). I also renamed `onTileCollision` to `regionCollidesWithATile`.

I fixed a bug in `clipPositionToBounds` where the ranges were not being compared to the right number. Its use of `min` and `max` also redundantly compared the number to itself, so I removed that.

I capitalized words in the middle of variable names in Perception to make them easier to read.